Fix gallery css, clean up gallery code, export Gallery and use separate index page for app DOM rendering
