+++ client/src/GalleryImage.jsx
@@ -4,6 +4,7 @@ import gallery from './gallery.css';
 const GalleryImage = ({image, length, toggle}) => {
   let id = image.img_order;
   const imgClass = `img${id}in${length}`;
+  console.log(imgClass);
   const url = `url(${image.url})`;
   return (
     <div className={gallery[imgClass]} style={{ backgroundImage: url }} id={id}>  </div>
