+++ client/src/LightboxComponents/lightbox.css
@@ -1,13 +1,21 @@
-.hide {
-  display: none;
-}
-
-.show {
+.modal {
   display: flex;
   align-items: center;
+  transition: 1.1s ease-out;
+  filter: blur(0);
+  transform: scale(1);
+  opacity: 1;
   position: absolute;
   background-color: white;
   height: 100%;
   width: 100%;
   z-index: 1;
+}
+
+.modal .off {
+  opacity: 0;
+  display: none !important;
+  visibility: hidden;
+  filter: blur(8px);
+  transform: scale(0.33);
 }
